fixes #245 : broker shouldn't allow clients to publish to '$' topics
